update meeting every minute

+++ database/migrations/2021_09_06_150622_create_transaction_table.php
@@ -28,13 +28,13 @@ class CreateTransactionTable extends Migration
             $table->string('shipping_name')->nullable();
             $table->text('shipping_address_name')->nullable();
             $table->string('shipping_address')->nullable();
-            $table->integer('shipping_province_id')->default(0);
+            $table->unsignedBigInteger('shipping_province_id')->default(0);
             $table->string('shipping_province_name')->nullable();
-            $table->integer('shipping_city_id')->default(0);
+            $table->unsignedBigInteger('shipping_city_id')->default(0);
             $table->string('shipping_city_name')->nullable();
-            $table->integer('shipping_district_id')->default(0);
+            $table->unsignedBigInteger('shipping_district_id')->default(0);
             $table->string('shipping_district_name')->nullable();
-            $table->integer('shipping_subdistrict_id')->default(0);
+            $table->unsignedBigInteger('shipping_subdistrict_id')->default(0);
             $table->string('shipping_subdistrict_name')->nullable();
             $table->string('shipping_zipcode')->nullable();
             $table->decimal('shipping_price', 26, 2)->nullable();
